pyproject.toml
[project]
name = "anticorruption"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "google-genai>=1.73.1",
    "pymupdf>=1.27.2.3",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "streamlit>=1.56.0",
    "uvicorn>=0.46.0",
]
